Improve SQL parsing performance
During our benchmark tests, we found out that our process spent 90% of its time
in the routine {ECLI_SQL_PARSER}.parse. This is mainly due to the fact that we
connect and disconnect from the database at each (HTTP) request. New statements
are therefore created and the SQL must be parsed again, even if it has already
been parsed by a previous (HTTP) request.

This patch implements a simple SQL parsing cache. We found out that this
reduces the time spent in {ECLI_SQL_PARSER}.parse to just 1% (tests were run
100 times to minimize the boostrap phase).